Data.SpreadSheet.Date

Description

Utilidades para poder trabajar mejor con fechas. No se pretende sobreescribir o hacer un nexo con librerías como Time, pero sí ofrecer algunas funciones que pueden ser comunmente utilizadas en hojas de cálculo.

Documentation

data Month

Enumerable con los meses.

Constructors

January

February

March

April

May

June

July

August

September

October

November

December

data WeekDay

Enumerable con los días de la semana.

Constructors

Monday

Tuesday

Wednesday

Thursday

Friday

Saturday

Sunday

next :: (Eq a, Bounded a, Enum a) => a -> a

Permite iterar sobre datos enumerables de forma circular. Una vez se llega al último elemento, el siguiente será el primero.

>>> next October
November
>>> next December
January

Permite la generacion de una lista infinita:

>>> take 11 $ iterate next Saturday
[Saturday,Sunday,Monday,Tuesday,Wednesday,Thursday,Friday,Saturday,Sunday,Monday,Tuesday]

getMonth :: Day -> Month

Indica el mes al que pertenece un día.

>>> getMonth $ fromGregorian 2019 1 14
January

Orphan instances

Arbitrary Day

Instance details

Methods

arbitrary :: Gen Day

shrink :: Day -> [Day]