Data.SpreadSheet.Internal.Gen

Description

Conjunto de generadore e instancias de Arbitrary para QuickCheck. Útiles para el ejemplo y para los módulos de test.

Documentation

genDay :: Gen Day

Generador de fechas próximas al final de 2018.

Orphan instances

Arbitrary Natural
Instancia para los naturales, igual que para Integer pero solo de sus valores absolutos.

Instance details

Methods

arbitrary :: Gen Natural

shrink :: Natural -> [Natural]

Arbitrary Day
Usando genDay.

Instance details

Methods

arbitrary :: Gen Day

shrink :: Day -> [Day]

Arbitrary Range
Se deriva la instancia de Natural

Instance details

Methods

arbitrary :: Gen Range

shrink :: Range -> [Range]

Arbitrary Cell
Distribuye de forma equitativa la generación de valores: - CNumber - CString - CBool - CDay

Instance details

Methods

arbitrary :: Gen Cell

shrink :: Cell -> [Cell]

CoArbitrary Natural
Tomada a partir de los Integer

Instance details

Methods

coarbitrary :: Natural -> Gen b -> Gen b

Function Natural
Se obtiene a través de los Integer

Instance details

Methods

function :: (Natural -> b) -> Natural :-> b

Arbitrary a => Arbitrary (SpreadSheet a)
Se deriva la instancia de [(Pos, a)]

Instance details

Methods

arbitrary :: Gen (SpreadSheet a)

shrink :: SpreadSheet a -> [SpreadSheet a]

CoArbitrary a => CoArbitrary (SpreadSheet a)
Tomada a partir de la lista de valores. [a]

Instance details

Methods

coarbitrary :: SpreadSheet a -> Gen b -> Gen b

Function a => Function (SpreadSheet a)
Tomada a partir de las listas de valores con posiciones.

Instance details

Methods

function :: (SpreadSheet a -> b) -> SpreadSheet a :-> b